chore: restructure hooks folder
