+++ src/app/qrcode/qr-url/page.tsx
@@ -7,8 +7,7 @@ import CategoryDropdownItems from '@/components/category-dropdown-item';
 import Input from '@/components/input';
 import ShortenCategories from '@/components/shorten-categories';
 import AuthContext from '@/contexts/authContext';
-import useAuthRouter from '@/hooks/useAuthRouter';
-import useScreenSize from '@/hooks/useScreenSize';
+import { useAuthRouter, useScreenSize } from '@/hooks';
 import meService from '@/libs/api/me';
 import categoryService from '@/services/category.service';
 import organizationService from '@/services/organization.service';
