+++ src/app/shorten/page.tsx
@@ -11,8 +11,7 @@ import ModalShorten from '@/components/modal-shorten';
 import ShortenCategories from '@/components/shorten-categories';
 import ShortenTools from '@/components/shorten-tools';
 import AuthContext from '@/contexts/authContext';
-import useAuthRouter from '@/hooks/useAuthRouter';
-import useScreenSize from '@/hooks/useScreenSize';
+import { useAuthRouter, useScreenSize } from '@/hooks';
 import meService from '@/libs/api/me';
 import categoryService from '@/services/category.service';
 import organizationService from '@/services/organization.service';
