+++ src/components/auth/auth-form/index.tsx
@@ -3,8 +3,7 @@ import { useState } from 'react';
 import Button from '@/components/button';
 import Input from '@/components/input';
 import InputCode from '@/components/input-code';
-import useAuthRouter from '@/hooks/useAuthRouter';
-import useEventListener from '@/hooks/useEventListener';
+import { useAuthRouter, useEventListener } from '@/hooks';
 import AuthFormFieldEnum from '@/types/auth-form-field-enum';
 import AuthFormFieldType from '@/types/auth-form-field-type';
 import AuthType from '@/types/auth-type-enum';
