+++ src/components/auth/auth-modal/index.tsx
@@ -3,8 +3,7 @@ import { useContext, useEffect, useState } from 'react';
 import CloseButton from '@/components/close-button';
 import AuthContext from '@/contexts/authContext';
 import AuthFormContext from '@/contexts/authFormContext';
-import useAuthRouter from '@/hooks/useAuthRouter';
-import useEventListener from '@/hooks/useEventListener';
+import { useAuthRouter, useEventListener } from '@/hooks';
 import authHeaderContent from '@/libs/auth-header-content';
 import AuthType from '@/types/auth-type-enum';
 import { detectOS } from '@/utils/common';
