+++ src/components/auth/login-common/index.tsx
@@ -4,8 +4,7 @@ import { useState, useContext, useEffect } from 'react';
 import ModalAlert from '@/components/modal-alert';
 import AuthContext from '@/contexts/authContext';
 import AuthFormContext from '@/contexts/authFormContext';
-import useAuthRouter from '@/hooks/useAuthRouter';
-import useInputErrorText from '@/hooks/useInputErrorText';
+import { useAuthRouter, useInputErrorText } from '@/hooks';
 import AlertLevel from '@/types/alert-level-enum';
 import AuthFormFieldEnum from '@/types/auth-form-field-enum';
 import AuthType from '@/types/auth-type-enum';
