+++ src/components/modal-alert/index.tsx
@@ -4,7 +4,7 @@ import AOS from 'aos';
 import Image from 'next/image';
 import { useEffect } from 'react';
 
-import useEventListener from '@/hooks/useEventListener';
+import { useEventListener } from '@/hooks';
 import AlertLevel from '@/types/alert-level-enum';
 
 import Button from '../button';
